fix(qsx): Fix signed/unsigned integer comparison warning -- src/qp/qsx/qsxUtil.cpp
Analysis:

1. Line 529 validates that the original key > 0: IDE_TEST_RAISE( *(SInt *)sKeyPtr <= 0, ERR_ARRAY_INDEX_OUT_OF_RANGE )
2. Line 536 assigns the key value to sKey (guaranteed > 0)
3. Line 537 performs sKey = sKey - 1, resulting in a value >= 0
4. By line 608, sKey is always >= 0, therefore (sKey + 1) is always > 0

So, changed line 608 from "if ( sCount != (sKey + 1) )" to "if ( sCount != (UInt)(sKey + 1) )".

Note:
[ fix warnings ]

src/qp/qsx/qsxUtil.cpp:608:21: warning: comparison of integers of different signs: 'UInt' (aka 'unsigned int') and 'int' [-Wsign-compare]
  608 |         if ( sCount != (sKey + 1) )
      |              ~~~~~~ ^   ~~~~~~~~

+++ src/qp/qsx/qsxUtil.cpp
@@ -605,7 +605,7 @@ IDE_RC qsxUtil::assignValue (
             sPrevKey = sKey;
         }
      
-        if ( sCount != (sKey + 1) )
+        if ( sCount != (UInt)(sKey + 1) )
         {
             sHasNull = 1;
         }
